int result

+++ ft_atoi.c
@@ -6,7 +6,7 @@
 /*   By: tkhemniw <gt.khemniwat@gmail.com>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/08/25 15:28:52 by tkhemniw          #+#    #+#             */
-/*   Updated: 2022/09/01 01:26:26 by tkhemniw         ###   ########.fr       */
+/*   Updated: 2022/09/01 01:32:46 by tkhemniw         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,9 +14,9 @@
 
 int	ft_atoi(const char *str)
 {
-	int			i;
-	int			sign;
-	long long	result;
+	int	i;
+	int	sign;
+	int	result;
 
 	if (!str)
 		return (0);
